src/models/requirements.txt
xgboost
lightgbm
catboost
scikit-learn
imbalanced-learn
numpy
pandas
mlflow
boto3
python-dotenv
pyyaml
tqdm
tqdm-joblib
